Mark vixl-test-runner as test, fix run target
With vixl-test-runner defined as a test, we can attach its test data to
it, so that everything needed to run the test is installed to a single
directory.

It will shortly become an error to read ANDROID_HOST_OUT from make,
since it's not guaranteed to be available. With vixl-test-runner fully
installed, we don't need to do all the path operations that were
previously necessary.

This also switches the test run target to use a timestamp file, so that
it's only rerun if the underlying test changes.

Test: m run-vixl-tests
Change-Id: If1bae1531cc89e4132d6b32ce227378c07a2abf0