Setup PGO for vixl
Bug: http://b/63768402

Setup PGO for libvixl-arm.so and libvixl-arm64.so.  The setup for
dex2oat and related libraries is done in an accompanying CL for that
project.  An instrumented copy of dex2oat can be produced for the device
using:
  ANDROID_PGO_INSTRUMENT=dex2oat make

Profiles for the instrumented libraries can be produced by running them
on a representative workload.

Currently, profile use is disabled (via enable_profile_use: false).
This can be turned on once a profile is checked in.

Test: Build with ANDROID_PGO_INSTRUMENT=dex2oat and verify profiles get
generated.

Change-Id: I00ef74b5f4f5deec84edcdc2e7d2a29698c9b546